filter-by-genre-feature

+++ src/main/java/com/microsoft/cse/helium/app/utils/ParameterValidator.java
@@ -100,7 +100,7 @@ public class ParameterValidator {
   public Boolean isValidRating(String rating) {
     if (!StringUtils.isEmpty(rating)) {
       try {
-        Integer movieRating = Integer.parseInt(rating);
+        Double movieRating = Double.parseDouble(rating);
         if (movieRating < 0 || movieRating > 10) {
           return false;
         }
